requirements.txt
python-telegram-bot==22.2
flask==3.1.1
requests==2.32.4
aiohttp==3.12.14

